feat(widgets): add `$$type` to widgets definition (#3960)
Adding a type to the `index` widget is necessary to filter it out from the list of the widgets. This is needed to implement the "reset page" behavior to get all the `index`-type children and reset their own page.

I took this opportunity to add a type to all widgets for debugging purpose.

The `$$type` property is not required in TypeScript because once we expose the widget interface, users might not need to mark them.