feat(connectMenu): update getWidgetSearchParameters (#4054)
This PR updates the lifecycle `getWidgetSearchParameters` in `connectMenu` to add the configuration of facets and default value. We don't apply the previous value on purpose, the `uiState` drives completely the `SearchParameters`.

I've not removed the `getConfiguration` lifecycle (yet) to avoid breaking everything. Once we've made the switch to `getWidgetSearchParameters` we'll remove them.