feat(connectHierarchicalMenu): update getWidgetSearchParameters (#4053)
This PR updates the lifecycle `getWidgetSearchParameters` in `connectHierarchicalMenu` to add the configuration of facets and default value. We don't apply the previous value on purpose, the `uiState` drives completely the `SearchParameters`.

I've not removed the `getConfiguration` lifecycle (yet) to avoid breaking everything. Once we've made the switch to `getWidgetSearchParameters` we'll remove them.