chore(configure): participate in routing (#4056)
* chore(configure): participate in routing

IFW-944

this adds simple routing for configure, where it will override the state on getWidgetState, and merge current state with configuration on getWidgetSearchParameters

This also changes the behaviour of the simple state mapping to explicitly filter out configure. While we don't expect someone to explicitly add a configure to the route, we can't be sure, and people need to explicitly override the statemapping to identity functions if they want to keep it. We can document that in the routing API documentation once we have it.

* chore: clarify comment

* feat: merge existing configuration on getWidgetState

* test: solve question

* Update src/lib/stateMappings/simple.ts

* Update src/connectors/configure/connectConfigure.ts

Co-Authored-By: Samuel Vaillant <samuel.vllnt@gmail.com>