feat(index): use uiState driven SearchParameters (#4059)
**Summary**

This PR replaces the usage of `getConfiguration` for `getWidgetSearchParameters`. The widgets are now driven by the `uiState`. Note that widgets that don't implement the lifecycle can still write on `SearchParameters`, but if a widget that implements the lifecycle use the same parameter it will be always overridden.

The lifecycle is now called on `init`, `add` (when the instance is started), `remove` (when the instance is started). On `init` we currently always provide an empty `uiState` (but it will change soon). On `add`/`remove` we provide the local `uiState` of the index to drive the `SearchParameters`.