fix(configure): merge with the previous parameters (#4085)
This PR uses the `mergeSearchParameters` for `getWidgetSearchParameters` within `configure`. With the changes introduced recently, we don't merge the parameters auto-magically. This logic is now managed by each widget itself. For `configure` this behavior was handy because it allowed us to use it for initial refinement. 

It's not possible anymore because we use `setQueryParameters` which overrides the previous parameters. The facets configuration of the current widget is erased and we're not able to refine anymore. To ease the transition between the two versions and keep the widget powerful enough we'll mimic the previous implementation.