refactor(router): remove widget dependency (#4120)
This PR removes the dependency of `Widget` on the type `Router`. The only method that both have in common is `dispose` but with a different signature. The `Router` doesn't have to receive `Helper`, `State`. The only data structures it has to care about are: `UiState`, `RouteState`. 

The update of the signature is indeed a breaking change but it shouldn't have an impact on user-land. The implementation of a custom router is rare. The usage of the arguments of `dispose` would be even rarer.