refactor(InstantSearch): RoutingManager integration (#4122)
This updates how the `RoutingManager` and `InstantSearch` interact together. 

`InstantSearch` is aware of `stateMapping` and `RouteState` which should not be the case. The only part that is concerned by those details is `RoutingManager`. The `InstantSearch` instance only cares about `UiState`. The `write` part of the  `router` is correctly abstracted behind a `write` method. We have now a `read` function exposed by the manager to avoid manipulate `router` and `stateMapping`.

The subscription to `onUpdate` function is a side effect of `RoutingManager` constructor. While it might be ok, it blocks us to move the declaration of the manager earlier inside the constructor. We already subscribe to the router a bit earlier than before but we don't want to subscribe if the instance does not start. We have now a `subscribe` function exposed by the manager that triggers the subscription to `onUpdate`.