fix(history): avoid empty query string (#4130)
**Summary**

This PR fixes an issue with the `historyRouter` + `simpleStateMapping`. They push an empty query string into the URL. The problem arise because we now always have an object with at least on key which is the top level index. The condition to add or not the query string is done on the `routeState` but it could be done on the query string itself since we already have computed it.

**Before**

![before](https://user-images.githubusercontent.com/6513513/65231669-2f1d1a80-dad0-11e9-8184-bd0ab3a8430d.gif)

**After**

![after](https://user-images.githubusercontent.com/6513513/65231680-32b0a180-dad0-11e9-9928-6a5231dfee60.gif)