feat(autocomplete): provide indexId (#4142)
This PR adds the `indexId` attribute to the `indices` provided by `autocomplete`.

Only the `indexName` was provided but its value is not stable. The value changes when the index changes e.g. with `sortBy`. A possible use case for stable value is a list of labels. Users might want to have labels on their indices. Without stable values, you can't predict which label you'll have.  The `indexId` solve this problem.