fix(helper): expose .lastResults to .helper
In v3, the helper exposed on the instance is the same helper as the one who searches. This means that lastResults is set as expected.

However, with v4 we split this up in the derived helper which actually searches and the exposed helper.

This is fixed now by just setting the last results once the search is done on the derived helper.