fix: widget name in documentation link for index
For technical reason we were not able to use `index` as widget name in
the docs, so we picked instead `index-widget`.

When you misuse a widget, we output an error message with a link to the
documentation generated from the name of the widget.

This commit handle the special case of the index widget.