feat(transformers): add tests (#4153)
* feat(transformers): add CLI

This CLI is mostly copied from react-codemod, which is BSD licensed, so I need to look at where to put the license text.

Since the CLI has dependencies, which I just added in teh dependencies field, it might make sense to have a separate repo. I like the fact that you can run it from the dependencies directly though, so to consider.

While testing this out I found a few more changes, which were applied in the next commit.

* chore: additional changes found by fix in script

* Update addWidget-to-addWidgets.test.js

* Update scripts/transforms/README.md

* chore: add license

* chore: move dependency

* chore: add to package.json

* chore: replace chalk with inline

* fix typo

* Update scripts/transforms/README.md

Co-Authored-By: François Chalifour <francoischalifour@users.noreply.github.com>

* move file structure less complex

* Update scripts/transforms/README.md

* remove CLI

* chore: ordering

* docs: wording of call