fix(insights): move 'insightsClient not provided error' to wrapper level
Prior to this change, this error was fired in the click listeners. As a result, users who used a custom component (and therefore a custom listeners) could never see it (they saw instead: insights is `TypeError: insights is not a function`).

Now the insightsClient wrapper itself is responsible for throwing this error.

This also fixes a bug introduced in https://github.com/algolia/instantsearch.js/pull/4197 when the `if (!props.insights) { ... }` was accidentaly brought on top of the handler.

This is not breaking change.