feat(helper): forward events from indices
This fixes a confusion when you want to listen to helper events (like change and result), but they aren't fired, since the root helper is not the actual one searching (the extra helper in the index is).

I'm making the PR so it can be found back in the future, but I don't think we should go for this now, since it encourages helper usage.