WIP: pattern router
This is a proof of concept of a new router, its advantage is that it's way more readable than a classic SEO history router. I will write an RFC on this once there's a first batch of rough feedback.

The example usage is in the media example, so that

Things to think about:

- how do we deal with this dependency?
- is this something we want to offer
- technically: clean up the "omit" part for which params go to query string
- do people want more control over e.g. hash parameters?
- do we need a polyfill of URL in the docs, or do we rewrite it to not need it?
- naming?  we can't both call the parameter "pattern", as well as the router
- does this become an option in historyRouter instead of a separate router?