feat(highlight): add cssClasses to snippet & highlight helper (#4306)
* feat(highlight): add cssClasses to snippet & highlight helper

This is related to: https://github.com/algolia/vue-instantsearch/issues/759. Just adding this doesn't totally solve the situation yet, but it at least allows a hook for adding extra class names.

What's interesting to note is that the suit helper in InstantSearch only concatenates names, and thus the keys for cssClasses are just the modifier. In Vue InstantSearch the keys for classNames are the generated classes, since the suit helper will check those keys directly. I wonder if we could move to such a system in InstantSearch.js in the future. This will allow us to remove these mappings, and deal with it at call site (which likely has a decrease in bundle size)

* add comment explaining why  helper isn't used

* chore: change bundlesize slightly