fix(router): merge results of intialUiState with RouteState
Initial UI State can be used to set a default refinement, e.g. an item in a refinement list to be checked. This can be done in parallel to routing, and should merge per index.

Slightly changed the code so that typescript didn't think the instance was "any"

related to #4393, but not fixing it, since they want intial UI State to be _written_ to the router too, which probably shouldn't be done, since that will cause a navigation as an effect of loading

Before this PR the second new test was failing