chore(deps): remove eslint-import-resolver-webpack
I've done a little tour through git history to find when it was added (a part of eslint-config-algolia, in 2016 originally as far as I can see), but it does not seem to be used anywhere anymore now.

I originally had the idea to remove this package to remove the dependency on "elliptic", but webpack still depends on it. At least now the dependencies are slightly simpler!