fix(insightsMiddleware): throw an error when credentials can't be extracted
When api or appId is missing, an error gets thrown by search insights, but that doesn't explain *why* that error gets thrown.

Ideally this would also clarify what the next step is (usually making sure to spread the previous search client, or leaving all properties in place, but i'm not too sure how to word that.

This doesn't add a new error where there wasn't already one, as the search-insights library threw an error in this case already.