feat(dynamicWidgets): implement additionalParameters
The default additionalParameters of `{ facets: ['*'], maxValuesPerFacet: 20 }` prevents an extra network request when widgets mount/unmount, as both queries will be identical, and thus cached.

If a refinement is set, this still results in two network requests, as the refinements can't be applied before it's known if the widget should be mounted.

todo: write the tests