fix(dynamicWidgets): allow only dynamicWidgets as attribute
According to: https://handlebarsjs.com/guide/builtin-helpers.html#if, an empty array will not render the block at all. If there's no attributesForFaceting, but there's dynamicWidgets, the sidebar should still render.

Since this change the sidebar is always rendered, but that's no big deal as the css is always there too