feat(DynamicWidgets): throw when transformItems returns anything that isn't an array
`transformItems` implementations can get complicated and if you're not
cautious, it's very easy to have it return something other than an
array.

When this happens, the errors you get are not insightful on where the
mistake is.

```
TypeError: Cannot read property 'forEach' of undefined
```

This change helps the user identify this type of mistakes.
It adds a check on `transformItems` output to make sure it returns an
array.

closes FX-608