fix(events): emit error as typeof Error (#4983)
* fix(events): emit error as typeof Error

If an uncaught error is thrown (eg. from the search engine) we want to throw that as is and not `Uncaught, unspecified \"error\" event. ([object Object])`

* write a test

couldn't write one with the uncaught error, as you can't test that an error is uncaught (not even by a try catch, as it happens async