feat(refinements): escape facet values starting with "-"
If a facet starts with `-`, the filter will be equivalent to "NOT VALUE" instead of "-VALUE". This is fixed by using the "value" key of each facet instead of the name, via https://github.com/algolia/algoliasearch-helper-js/pull/889

This refactored some, but not all of the tests (as all would be too much work) to follow a more realistic code path.

FX-1135