fix(routing): prevent writing the same URL twice
We've given this from time to time as workarounds for customer issues, where due to state desync or other external influence the URL is already changed to the one InstantSearch would navigate to.

The specific state I test and am fixing here however is with `stateMapping` (described in https://github.com/algolia/instantsearch.js/discussions/5044) removing a piece of the state that still changes, and thus would cause duplicate writes of the same URL before this PR.

I had to change the test for replaceState a bit, as the cleaned URL being the same as the replaceState navigated URL being the same was hiding the issue we have that test for.