fix(status): send render event when loading starts
Without this it's impossible to use `loading` for anything. I opted to only emit the event, as calling render on the widgets would  require changes in all the tests.

If we render (fully), you still don't get updated of `loading` initially because `render` only gets called when there are results:

https://github.com/algolia/instantsearch.js/blob/4d43bd8292446101742a10d1da9f042f780d8d45/src/widgets/index/index.ts#L573-L575

To fix all of these inconsistencies, we should join `init` and `render` in a single function (always calling render, even if it's the first render, also when there's no results), but this would be a significant breaking change, so shouldn't be done lightly