tests(unit): delay "wait" for more stability
It seems that in some cases the unit tests which wait for a timeout to "definitely have happened" don't wait long enough. AFAICT this is because setTimeout doesn't have an implicit ordering, if the timer in the code is delayed (setTimeout argument is a minimum), it's possible that the timer in the test already has expired, and thus trying to assert too fast.

I could not find other patterns to replace these type of tests with, so added a decent chunk of extra timeout, hoping that these tests from now on no longer are flaky!