test(jest): avoid flaky modules test
The modules-* test about dependency detection of react-dom in react-instantsearch-hooks is flaky with regular jest workers, as I believe sometimes the jest mock of a module becomes global and still relevant for the next test.

Ideally we wouldn't need to do such tests when react-dom is exclusively user-provided, but for now hopefully running the tests in band will avoid that test being flaky.