fix(getServerState): correctly forward asynchronous errors
fixes https://github.com/algolia/instantsearch/issues/5233

What happens is that a synchronous error (as we had in the test, but not really in real life) gets caught by the derived helper as it happens before the response, but an async (regular) error happens in the response parsing, and thus before the derived helper even receives the result.

It gets rethrown by InstantSearch, but we also catch the errors in the main helper, just in case there's another case not covered here.