fix(url): add widgets early always (WIP)
see: https://github.com/algolia/instantsearch/issues/5236

This should fix it in theory, although i can only reproduce this in real-life situations so far. Tests aren't yet adjusted.

As shouldSsr is set to true in connectors, it probably also be set to true in useConnector, and therefore probably just set to true always. It is slightly more expensive as the widget is checked in the list first.