fix(dependencies): update helper requirement for minor PP vulnerability
The vulnerability comes up when user-defined search parameters have a malicious key `constructor`. An error then will be thrown, so the search will no longer be active, but any prototype injection would still happen, causing people to be vulnerable.

With default implementations of InstantSearch, this isn't reachable, as `searchParameters` aren't a direct parameter, except for `configure`