fix(insights): pass correct sendEvent to html template
There was an existing prop called `sendEvent`, which we passed through to the component when we introduced html templates. However, this sendEvent was actually sendEventToInsights, which is of course problematic, as it doesn't match the documentation, nor does it add the necessary conversion from a simple "three argument" form to "object form".

This PR changes the prop that existed to `sendEventToInsights`, as well as passing down the `sendEvent` which comes from the connector.

discovered in https://github.com/algolia/instantsearch/discussions/5447#discussioncomment-4779169