fix(infiniteHits): read cache correctly when search is loading
The cache is based on search parameters (minus page), so when the state updates, we want to still use the previous state, as the previous hits are still valid.

One thing that could change this design is if the cache wasn't purely based on the parameters, and somehow was persisted too, so we could ask for the "last result", but that's not possible without major change in the cache.