fix(HitsPerPage): compute selected from isRefined
This makes it responsive to the URL, as expected like other widgets.

fixes https://github.com/algolia/instantsearch/issues/5397

Slight improvement in the HitsPerPage + SortBy tests, as it was relying on Vue state instead of the DOM

I wanted both to be exactly the same, but SortBy doesn't have isRefined on the items apparently.