fix(insights): pass credentials in payload
If something else controls insights client too, it could be initialised with the wrong parameters. To ensure the right app is targeted, we pass appId and apiKey in the payload as well.

This requires https://github.com/algolia/search-insights.js/pull/398 or similar to have any effect though

also for onEvent