fix(rangeinput): allow input of numbers with precision
While we set the `step` correctly based on the precision, we didn't actually allow input of values with decimal separator because the "inbetween" situation (like 1.) isn't a valid number, causing the state to reset.

To solve this, we prevent saving the value to react state if the value is invalid. The browser will take over from there and show warnings if it's not a number.

CR-2948