refactor: move dedupe into middleware
this allows the dedupe to work for both bindEvent and sendEvent in a single codebase.

Tiny aside: this means that the "next internal click" is now global, meaning if you click on geo, then on hits, we'll see that as an internal click to prevent sending. This is an extreme edge case, as almost nowhere sends click events.