feat(insights): only automatically add middleware based on condition
similar to #5557, but for the whole middleware.

FX-2292

(this obviously didn't update the tests, as the effect is widespread, mainly no userToken for the first search)

Another thing to figure out is that `clickAnalytics` isn't set to true, so that first request doesn't have a queryID. Maybe that should be fixed backend (to also send one if not asked)?