refactor: centralise code for SSR in InstantSearch.js
A couple utilities were rewritten in Vue and React Hooks, to be able to make SSR. Moving these utilities into InstantSearch.js makes it so they don't need to be rewritten when we add SSR to JS or another flavor.

Note a little fix to `walkIndex` as well, it used to call every index (except the first one) two times.