fix(removeWidgets): defer search by a full tick
In React InstantSearch Hooks, the removeWidget cleanup is deferred by a timeout, as it ensures the widget doesn't get recreated in strict mode needlessly.

However, this meant that each one of the removeWidget calls happened in their own context, as the order was

timeout (remove)
promise (defer)
timeout (remove)
promise (defer)

and thus was never deduplicated.

Still to figure out in this PR whether the scheduleSearch using a timer is the way to go, or if `scheduleSearch` in index itself should be in a setTimeout (did not yet verify whether that would work too).

I'd also like to know if the order in which different timers are fired is consistent enough for this to work always. We may need to reimplement a queue ourselves.