fix(lifecycle): prevent extra network requests when unmounting multiple widgets
When unmounting multiple widgets at once, but not the InstantSearch widget, multiple searches were triggered. This was because every timeout leaves a moment for microtasks to be processed, and thus did not deduplicate scheduleSearch.

This PR fixes that by adding a private timeout-based scheduler that ensures all removeWidgets calls get called synchronously.

CR-3255