chore(connectToggleRefinement): remove duplicated types (#5690)
Types seems to have existed both in a types file and in the connector. I'm assuming this is an oversight from the TypeScript migration where it was migrated two times.

I did not remove the `types.ts` file in case someone depends on it, but can be removed in a next major. I did remove the duplicated types and re-export from the main file instead.