fix(server): allow detached index (no rootIndexName) in SSR
When you don't pass a root indexName, no search gets done for the root, therefore the searchResults are also `null` for that index.

With this fix, getInitialResults is resilient for this behaviour and ensures results are still returned as expected.

This also changes the error from an implicit "can't read _state from results" when there are no results and you're trying to read the results, to an explicit error.

fixes #5831