fix(createURL): ensure refined value gets removed
If there's only one refinement left, we used to keep `uiState` purposely as-is, although actually it needs to be cleared from the refinements.

This also has an impact (positive, removing stray refinements) on dispose with the new disposeMode likely, but to investigate.

This also needs to be done and tested for all other refinement widgets