feat(nextjs): prevent onUpdate/onStateChange on own write
Essentially, because we use router.push to write the URLs, we can't distinguish from a third-party URL/route update, which obviously should be taken in account for InstantSearch compared to a push we do ourselves in push/write, of which InstantSearch is already aware.

Impact this would be preventing is double `onStateChange` calls when you're using next routing (CR-5006)

not yet written a test, so not sure if this works, or if it's even the right approach (maybe url should be checked instead of an ad-hoc variable).

As far as i can tell react-instantsearch-nextjs doesn't have this issue as it doesn't listen to the route change event (is that an oversight?)