fix(packaging): remove non-exposed "source" key
This is what currently causes the `createElement is not defined` error. Basically the example builds "from source", but something in the build (minification) renames `createElement` to `e` and thus everything breaks.

In this PR I remove the `source` key to ensure no building is happening from source, everything (mainly our examples) will use the built code.

This has no (except possibly positive) impact for customers, as in all these packages `src` is not published, so the source refers to a non-existing file.