chore(bundlesize): update limit
This seems to be needed because the location of the babel helpers needs to change to support this new usage of spread.

I don't see any rhyme or reason to why it actually happens or why that has such a large impact for just moved code, but the babel helpers clearly are already very inefficient (many duplicates of the same helper functions) that this probably is fixed if we take a look later to improve this.