fix(hits): rename the exposed "hits" to "items"
`hits` of course still works for now, but `items` is also available from the connectors.

A warning on getting of `hits` from the connector is given, which does not fire on normal usage of the components.

For now, no change is done in `ui-components`, but I'm happy to do so if we're happy with the hypothetical breaking change for that package.