fix(routing): manually push changed path
Instead of using next's `router.push` function, we're transitioning back to `history.pushState`. This works better in many edge cases as it prevents a whole page refresh and rerender.

Possible intended side effect of this fix is of course that any global change handlers or code running in render won't run anymore when changing instantsearch state. This can be worked around by using the `onStateChange` function or a middleware.

Back navigation still works as expected.